% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtls.R
\name{communicate}
\alias{communicate}
\title{URL building and request/response handling}
\usage{
communicate(q, corr, ld_method)
}
\arguments{
\item{q}{The qtlizer query. Can either be a single string or a vector.}

\item{corr}{Linkage disequilibrium based on 1000 Genomes Phase 3 European.
Optional value between 0 and 1. Default value is NA.}

\item{ld_method}{There are two methods. Default method is "r2". 
The other opportunity is to use "dprime".}
}
\value{
Data frame with results.
}
\description{
URL building and request/response handling
}
