% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtls.R
\name{mkQuery}
\alias{mkQuery}
\title{Actullaly makes the connection to server and returns results}
\usage{
mkQuery(q, corr = 0.8, ld_method = "r2")
}
\arguments{
\item{q}{The qtlizer query. Can either be a single string or a vector.}

\item{corr}{Linkage disequilibrium based on 1000Genomes Phase 3 European.
Optional value between 0.1 and 1. Default value is 0.8.}

\item{ld_method}{There are two methods. Default method is "r2". 
The other opportunity is to use "dprime".}
}
\value{
Data frame with response.
}
\description{
Actullaly makes the connection to server and returns results
}
