% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtFunction.R
\name{qtfun}
\alias{qtfun}
\title{Query Qtlizer}
\usage{
qtfun(query, outputFile = of_default)
}
\arguments{
\item{query}{The qtlizer query.}

\item{outputFile}{(optional) The name of the output file.}
}
\value{
Response goes directly into the output file.
}
\description{
Makes a query at Qtlizer and writes the results 
into a specified (or default) file.
}
\examples{
qtfun("rs4284742")
qtfun("rs4284742", "outputFile.txt")
}
