% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtls.R
\name{vector_split}
\alias{vector_split}
\title{Splits vector v into n subvectors}
\usage{
vector_split(v, n)
}
\arguments{
\item{v}{input vector}

\item{n}{number of subvectors}
}
\value{
List with subvectors.
}
\description{
Splits vector v into n subvectors
}
\keyword{internal}
