% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtls.R
\name{get_qtls}
\alias{get_qtls}
\title{Query Qtlizer}
\usage{
get_qtls(query)
}
\arguments{
\item{query}{The qtlizer query. Can either be a single string or a vector.}
}
\value{
Data frame with response.
}
\description{
Makes a query at Qtlizer and returns results as data frame.
}
\examples{
get_qtls("rs4284742")
get_qtls(c("rs4284742", "DEFA1"))

}
