% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtls.R
\name{get_qtls}
\alias{get_qtls}
\title{Query Qtlizer}
\usage{
get_qtls(query, r2 = 0.8)
}
\arguments{
\item{query}{The qtlizer query. Can either be a single string or a vector.}

\item{r2}{Linkage disequilibrium based on 1000Genomes Phase 3 European.
Optional value between 0 and 1. Default value is 0.8.}
}
\value{
Data frame with response.
}
\description{
Makes a query at Qtlizer and returns results as data frame.
}
\examples{
get_qtls("rs4284742")
get_qtls(c("rs4284742", "DEFA1"))
get_qtls("rs4284742", 0.6)
}
